Minimal Number of Coins for Change

+++ 0x0A-argc_argv/main.h
@@ -7,6 +7,6 @@
 #include <string.h>
 
 /* Declarations */
-long int get_coins(long int);
+int get_coins(int);
 
 #endif
